requirements.txt
openai
pdfplumber
re
logging
json
pdfminer
pdf2image
pytesseract
pillow
fitz
yaml
tqdm
mammoth
pydocx
markdownify
semantic-text-splitter
tokenizers